Refactor parts of the codebase.
