+++ bungeecord/src/main/java/com/github/koxsosen/bungee/BungeePluginLoader.java
@@ -1,8 +1,6 @@
 package com.github.koxsosen.bungee;
 
 import com.github.koxsosen.common.LibertyBansApiHelper;
-import com.github.koxsosen.bungee.event.PunishmentListener;
-import com.github.koxsosen.bungee.messaging.MessageReceiver;
 import net.md_5.bungee.api.plugin.Plugin;
 import space.arim.libertybans.api.LibertyBans;
 import space.arim.omnibus.Omnibus;
